Proyecto Biblioteca/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
diff-match-patch==20230430
Django==5.1.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-grappelli==4.0.1
django-import-export==4.1.1
django-modeltranslation==0.19.9
django-rosetta==0.10.1
django-simple-history==3.7.0
idna==3.10
mysqlclient==2.2.4
polib==1.2.0
pydot==3.0.1
pyparsing==3.1.4
requests==2.32.3
sqlparse==0.5.1
tablib==3.5.0
urllib3==2.2.3
